Split link utilities into separate files and limit re-exports
This should help make future refactoring a bit easier, and gets
us away from re-exporting all link utility functions, when very
few of them need to be made available outside of this project.
