+++ src/link/utils/__tests__/fromError.ts
@@ -1,0 +1,12 @@
+import { toPromise } from '../toPromise';
+import { fromError, } from '../fromError';
+
+describe('fromError', () => {
+  it('acts as error call', () => {
+    const error = new Error('I always error');
+    const observable = fromError(error);
+    return toPromise(observable)
+      .then(expect.fail)
+      .catch(actualError => expect(error).toEqual(actualError));
+  });
+});
