+++ src/link/utils/__tests__/fromPromise.ts
@@ -1,0 +1,25 @@
+import { fromPromise } from '../fromPromise';
+import { toPromise, } from '../toPromise';
+
+describe('fromPromise', () => {
+  const data = {
+    data: {
+      hello: 'world',
+    },
+  };
+  const error = new Error('I always error');
+
+  it('return next call as Promise resolution', () => {
+    const observable = fromPromise(Promise.resolve(data));
+    return toPromise(observable).then(result =>
+      expect(data).toEqual(result),
+    );
+  });
+
+  it('return Promise rejection as error call', () => {
+    const observable = fromPromise(Promise.reject(error));
+    return toPromise(observable)
+      .then(expect.fail)
+      .catch(actualError => expect(error).toEqual(actualError));
+  });
+});
