+++ src/link/utils/__tests__/validateOperation.ts
@@ -1,0 +1,17 @@
+import { validateOperation, } from '../validateOperation';
+
+describe('validateOperation', () => {
+  it('should throw when invalid field in operation', () => {
+    expect(() => validateOperation(<any>{ qwerty: '' })).toThrow();
+  });
+
+  it('should not throw when valid fields in operation', () => {
+    expect(() =>
+      validateOperation({
+        query: '1234',
+        context: {},
+        variables: {},
+      }),
+    ).not.toThrow();
+  });
+});
