+++ src/link/utils/fromPromise.ts
@@ -1,0 +1,12 @@
+import Observable from 'zen-observable';
+
+export function fromPromise<T>(promise: Promise<T>): Observable<T> {
+  return new Observable<T>(observer => {
+    promise
+      .then((value: T) => {
+        observer.next(value);
+        observer.complete();
+      })
+      .catch(observer.error.bind(observer));
+  });
+}
