+++ src/link/utils/toPromise.ts
@@ -1,0 +1,21 @@
+import Observable from 'zen-observable';
+import { invariant } from 'ts-invariant';
+
+export function toPromise<R>(observable: Observable<R>): Promise<R> {
+  let completed = false;
+  return new Promise<R>((resolve, reject) => {
+    observable.subscribe({
+      next: data => {
+        if (completed) {
+          invariant.warn(
+            `Promise Wrapper does not support multiple results from Observable`,
+          );
+        } else {
+          completed = true;
+          resolve(data);
+        }
+      },
+      error: reject,
+    });
+  });
+}
