+++ src/link/utils/validateOperation.ts
@@ -1,0 +1,20 @@
+import { InvariantError } from 'ts-invariant';
+
+import { GraphQLRequest } from '../core/types';
+
+export function validateOperation(operation: GraphQLRequest): GraphQLRequest {
+  const OPERATION_FIELDS = [
+    'query',
+    'operationName',
+    'variables',
+    'extensions',
+    'context',
+  ];
+  for (let key of Object.keys(operation)) {
+    if (OPERATION_FIELDS.indexOf(key) < 0) {
+      throw new InvariantError(`illegal argument: ${key}`);
+    }
+  }
+
+  return operation;
+}
